refactored GoBackToRoot
