+++ Source/Xamarin/Prism.Forms/Navigation/INavigateInternal.cs
@@ -7,6 +7,8 @@ namespace Prism.Navigation
     {
         Task<bool> GoBackInternal(NavigationParameters parameters, bool? useModalNavigation, bool animated);
 
+        Task GoBackToRootInternal(NavigationParameters parameters);
+
         Task NavigateInternal(string name, NavigationParameters parameters, bool? useModalNavigation, bool animated);
 
         Task NavigateInternal(Uri uri, NavigationParameters parameters, bool? useModalNavigation, bool animated);
