+++ Source/Xamarin/Prism.Forms/Navigation/INavigationServiceExtensions.cs
@@ -1,9 +1,5 @@
-﻿using Prism.Common;
-using System;
-using System.Collections.Generic;
-using System.Linq;
+﻿using System;
 using System.Threading.Tasks;
-using Xamarin.Forms;
 
 namespace Prism.Navigation
 {
@@ -27,46 +23,9 @@ namespace Prism.Navigation
         /// <param name="navigationService">The INavigatinService instance</param>
         /// <param name="parameters">The navigation parameters</param>
         /// <remarks>Only works when called from a View within a NavigationPage</remarks>
-        public static async Task GoBackToRootAsync(this INavigationService navigationService, NavigationParameters parameters = null)
+        public static Task GoBackToRootAsync(this INavigationService navigationService, NavigationParameters parameters = null)
         {
-            try
-            {
-                if (parameters == null)
-                    parameters = new NavigationParameters();
-
-                parameters.AddInternalParameter(KnownInternalParameters.NavigationMode, NavigationMode.Back);
-
-                IPageAware pageAware = (IPageAware)navigationService;
-
-                var canNavigate = await PageUtilities.CanNavigateAsync(pageAware.Page, parameters);
-                if (!canNavigate)
-                    return;
-
-                List<Page> pagesToDestroy = pageAware.Page.Navigation.NavigationStack.ToList(); // get all pages to destroy
-                pagesToDestroy.Reverse(); // destroy them in reverse order
-                var root = pagesToDestroy.Last();
-                pagesToDestroy.Remove(root); //don't destroy the root page
-
-                PageUtilities.OnNavigatingTo(root, parameters);
-
-                await pageAware.Page.Navigation.PopToRootAsync();
-
-                foreach (var destroyPage in pagesToDestroy)
-                {
-                    PageUtilities.OnNavigatedFrom(destroyPage, parameters);
-                    PageUtilities.DestroyPage(destroyPage);
-                }
-
-                PageUtilities.OnNavigatedTo(root, parameters);
-            }
-            catch (InvalidOperationException ex)
-            {
-                throw new InvalidOperationException("GoBackToRootAsync can only be called when the calling Page is within a NavigationPage.", ex);
-            }
-            catch
-            {
-                throw;
-            }
+            return ((INavigateInternal)navigationService).GoBackToRootInternal(parameters);
         }
 
         /// <summary>
